fix: App.jsx를 App.tsx로 변환 및 코드 정리
