+++ src/App.tsx
@@ -1,0 +1,15 @@
+import { ThemeProvider } from "styled-components";
+
+import GlobalStyles from "@/style/GlobalStyles";
+import { theme } from "@/style/theme";
+
+function App() {
+  return (
+    <ThemeProvider theme={theme}>
+      <div></div>
+      <GlobalStyles />
+    </ThemeProvider>
+  );
+}
+
+export default App;
